Finalised working autogg, also added better mod info in mcmod.info
