+++ src/main/resources/mcmod.info
@@ -1,11 +1,11 @@
 [
   {
     "modid": "${modid}",
-    "name": "Xample Mod",
-    "description": "A mod that is used as an example.",
+    "name": "AutoGGMod",
+    "description": "KanyeFan30's first Forge Mod",
     "version": "${version}",
     "mcversion": "${mcversion}",
-    "url": "https://github.com/romangraef/Forge1.8.9Template/",
+    "url": "https://github.com/KanyeFan30/autoggmod",
     "updateUrl": "",
     "authorList": [
       "You"
